Changed year from short to int; added Swagger documentation; changed endpoint to POST; moved claimant ID to route.

+++ DWD.UI.Monetary/DWD.UI.Monetary.Domain/BusinessEntities/ClaimantWage.cs
@@ -20,7 +20,7 @@ public partial class ClaimantWage
     /// <summary>
     /// Gets or sets wage Year.
     /// </summary>
-    public short? WageYear { get; set; }
+    public int? WageYear { get; set; }
 
     /// <summary>
     /// Gets or sets wage Quarter.
